Perf test for various function invocation methods.
This compares the result of calling Delegate.CreateDelegate, invoking Func<>.ctor directly, and simply calling the original delegate.

On my machine a very surprising result: just calling the original delegate is the *slowest*. Is this due to some defect in my test, or could it be that invoking a delegate with an explicit parameter is that much slower than calling a delegate with no explicit parameter?